# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Qubah Learning API",
        "description": "REST API for the Qubah Learning educational platform. Provides endpoints for authentication, subjects, topics, content, progress tracking, and parent dashboard.",
        "contact": {
            "name": "Qubah Team",
            "email": "support@qubah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get all subjects",
                "operationId": "2e23fc1b496972f57cce4c25358fa0ee",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Create a new subject",
                "operationId": "509ee5c1fad37a80431f9baafbee11e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "section_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "الرياضيات"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "مادة الرياضيات للصف الأول"
                                    },
                                    "section_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "thumbnail_path": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/subjects/{id}": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get a specific subject",
                "operationId": "e36deb6a61342b7ccf0beba9e26910c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Update a subject",
                "operationId": "35d27b75dbdcd2209374683f8f264591",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "section_id": {
                                        "type": "integer"
                                    },
                                    "thumbnail_path": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "order": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Subject not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Delete a subject",
                "operationId": "c1d70106535f4ec184fadd9247a71214",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Subject not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login via /api/v1/auth/login and use the returned access_token here.",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}